requirements.txt
google-generativeai>=0.3.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
wikipedia-api>=0.6.0
python-dotenv>=0.19.0
requests>=2.26.0
python-dateutil>=2.8.2
pywin32>=303; sys_platform == 'win32'
psutil>=5.9.0
PyAudio>=0.2.13
portaudio>=19.6.0
